+++ scripts/make-cpp-tests.py
@@ -22,19 +22,30 @@ myOS = platform.system()
 path = Path.cwd() / "build" / myOS / "cpp-tests"
 path.mkdir(parents=True, exist_ok=True)
 
+# Set the C++ compiler
+CppCompiler = "clang++" if myOS == "Windows" else "g++"
+
 # Base CMake command to build the python module
+cmakeCommand = f"cmake -G \"Unix Makefiles\" -B {path} -S ./tests-cpp -D PYBIND_LIB=OFF -DCMAKE_BUILD_TYPE={buildType} -DCMAKE_CXX_COMPILER={CppCompiler} -DSQLITECPP_RUN_CPPLINT=OFF -DLLVM_USE_CRT_DEBUG=MD -Dgtest_force_shared_crt=ON"
 if myOS == "Windows":
-    vcvarsallPath = "\"C:\\Program Files\\Microsoft Visual Studio\\2022\\Community\\VC\\Auxiliary\\Build\\vcvarsall.bat\""
-    os.system(f"{vcvarsallPath} x86_amd64")
+    cmakeCommand += ' -DCMAKE_MAKE_PROGRAM="C:/msys64/clang64/bin/mingw32-make.exe"'
 
-    clPath = "\"C:\\Program Files\\Microsoft Visual Studio\\2022\Community\\VC\\Tools\\MSVC\\14.34.31933\\bin\\Hostx64\\x64\\cl.exe\""
-    cmakeCommand = f"cmake -G \"Visual Studio 17 2022\" -A x64 -B {path} -S ./tests-cpp/ -DCMAKE_BUILD_TYPE={buildType} -DCMAKE_CXX_COMPILER={clPath} -DMAIALIB_VERSION_INFO=0.0.0"
-else:
-    cmakeCommand = f"cmake -G \"Unix Makefiles\" -B {path} -S ./tests-cpp/ -DCMAKE_BUILD_TYPE={buildType}"
+if (buildType == "Debug"):
+    cmakeCommand += " -DPROFILING=ON"
 
-# # If 'Windows' define the MinGW 'make.exe'
+# # Base CMake command to build the python module
 # if myOS == "Windows":
-#     cmakeCommand += ' -DCMAKE_MAKE_PROGRAM="C:/msys64/mingw64/bin/make.exe"'
+#     vcvarsallPath = "\"C:\\Program Files\\Microsoft Visual Studio\\2022\\Community\\VC\\Auxiliary\\Build\\vcvarsall.bat\""
+#     os.system(f"{vcvarsallPath} x86_amd64")
+
+#     clPath = "\"C:\\Program Files\\Microsoft Visual Studio\\2022\Community\\VC\\Tools\\MSVC\\14.34.31933\\bin\\Hostx64\\x64\\cl.exe\""
+#     cmakeCommand = f"cmake -G \"Visual Studio 17 2022\" -A x64 -B {path} -S ./tests-cpp/ -DCMAKE_BUILD_TYPE={buildType} -DCMAKE_CXX_COMPILER={clPath} -DMAIALIB_VERSION_INFO=0.0.0"
+# else:
+#     cmakeCommand = f"cmake -G \"Unix Makefiles\" -B {path} -S ./tests-cpp/ -DCMAKE_BUILD_TYPE={buildType}"
+
+# # # If 'Windows' define the MinGW 'make.exe'
+# # if myOS == "Windows":
+# #     cmakeCommand += ' -DCMAKE_MAKE_PROGRAM="C:/msys64/mingw64/bin/make.exe"'
 
 # Get CPU num threads
 numThreads = os.cpu_count()
@@ -42,13 +53,15 @@ numThreads = os.cpu_count()
 # Run CMake and Make commands
 os.system(cmakeCommand)
 
-if myOS == "Windows":
-    msbuildPath = "C:\\Program Files (x86)\\Microsoft Visual Studio\\2019\\BuildTools\\MSBuild\\Current\\Bin"
-    os.system(
-        f"msbuild.exe {path}\\cpp-tests.sln -verbosity:quiet -maxcpucount")
+os.system(f"make -j {numThreads} -C {path} --no-print-directory")
 
-# os.system(f "cmake --build {path} --config Debug")
-else:
-    os.system(f"make -j {numThreads} -C {path} --no-print-directory")
+# if myOS == "Windows":
+#     msbuildPath = "C:\\Program Files (x86)\\Microsoft Visual Studio\\2019\\BuildTools\\MSBuild\\Current\\Bin"
+#     os.system(
+#         f"msbuild.exe {path}\\cpp-tests.sln -verbosity:quiet -maxcpucount")
+
+# # os.system(f "cmake --build {path} --config Debug")
+# else:
+#     os.system(f"make -j {numThreads} -C {path} --no-print-directory")
 
 print(f"{color.OKGREEN}Build C++ Tests: Done!{color.ENDC}")